Size_t hash fix

+++ part2/array.h
@@ -87,7 +87,7 @@ public:
     /**
      * Returns the hash value for this Array.
      */
-    int hash()
+    size_t hash()
     {
     }
 };
